Added portable hash

+++ include/nanorpc/core/type.h
@@ -19,7 +19,7 @@
 namespace nanorpc::core::type
 {
 
-using id = std::size_t;
+using id = std::int64_t;
 using buffer = std::vector<std::uint8_t>;
 using executor = std::function<buffer (buffer)>;
 using executor_map = std::map<std::string, executor>;
